+++ src/main/java/nure/com/InbinFilter/dto/waste/WasteGetDto.java
@@ -8,8 +8,8 @@ public class WasteGetDto {
     private Long id;
     private String residentName;
     private LitterType litterType;
-    private Integer amount;
-    private Integer priceToPay;
+    private Double amount;
+    private Double priceToPay;
 
 
 }
